profile add/delete update
Adding a profile with other blank fields properly classifies those blank fields as NULL in the database. Clicking the 'X' button left of the FIRST SLOT deletes the profile from the database as intended.
(Please note: The deletion code is only applied to the first slot, the remaining slots need to be updated)